Move TestProfileService to dedicated file FakeProfileService
