+++ feature/login/src/commonMain/kotlin/remote/ProfileService.kt
@@ -11,7 +11,6 @@ import io.ktor.client.plugins.resources.get
 import io.ktor.client.plugins.resources.patch
 import io.ktor.client.plugins.resources.post
 import io.ktor.client.request.setBody
-import kotlinx.io.files.Path
 import org.pointyware.xyz.core.entities.Uuid
 import org.pointyware.xyz.core.entities.profile.DriverProfile
 import org.pointyware.xyz.core.entities.profile.RiderProfile
@@ -87,43 +86,5 @@ class KtorProfileService(
         } catch (e: Exception) {
             return Result.failure(e)
         }
-    }
-}
-
-/**
- * A test implementation of the [ProfileService] interface. Instead of relying on a remote service,
- * it will load and store profiles in a given directory between test runs.
- */
-class TestProfileService(
-    private val profileFile: Path,
-    private val profiles: MutableMap<Uuid, ProfileEntity> = mutableMapOf()
-): ProfileService {
-
-    // TODO: Implement file persistence for testing
-
-    override suspend fun createDriverProfile(userId: Uuid, profile: DriverProfile): Result<DriverProfile> {
-        profiles[userId] = profile
-        return Result.success(profile)
-    }
-
-    override suspend fun createRiderProfile(userId: Uuid, profile: RiderProfile): Result<RiderProfile> {
-        profiles[userId] = profile
-        return Result.success(profile)
-    }
-
-    override suspend fun getProfile(userId: Uuid): Result<ProfileEntity?> {
-        profiles[userId].let {
-            return Result.success(it)
-        }
-    }
-
-    override suspend fun updateProfile(userId: Uuid, profile: ProfileEntity): Result<ProfileEntity> {
-        profiles[userId] = profile
-        return Result.success(profile)
-    }
-
-    override suspend fun deleteProfile(userId: Uuid): Result<Unit> {
-        profiles.remove(userId)
-        return Result.success(Unit)
     }
 }
